# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.TableCreatable:
    properties:
      areaId:
        type: integer
      basePrice:
        type: number
      capacity:
        type: integer
      employeeId:
        type: integer
      normalTable:
        type: boolean
    type: object
  models.Warehouse:
    properties:
      categoryId:
        type: integer
      description:
        type: string
      discountPercent:
        type: integer
      minQuantity:
        type: integer
      name:
        type: string
      price:
        type: number
      stockAmount:
        type: integer
      thumb:
        type: string
      uint:
        type: string
    type: object
  models.response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      statusCode:
        type: integer
    type: object
host: localhost:3001
info:
  contact:
    email: dso.intern.xuanhoa@gmail.com
    name: Xuan Hoa Le
  description: Go Food Delivery CMS - Swagger API Discovery
  title: Go Food Delivery CMS - Swagger API Discovery
  version: "1.0"
paths:
...
  /tables:
    get:
      consumes:
      - application/json
      description: Read Table By Employee Id
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Read Table By Employee Id
      tags:
      - tables
    post:
      consumes:
      - application/json
      description: Create Table
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Table
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/models.TableCreatable'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Create Table
      tags:
      - tables
  /tables/all:
    get:
      consumes:
      - application/json
      description: Read All Table
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Read All Table
      tags:
      - tables
  /tables/status/{status}:
    get:
      consumes:
      - application/json
      description: Read Table By Employee Id And Status
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Status
        in: path
        name: status
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Read Table By Employee Id And Status
      tags:
      - tables
  /warehouses:
    patch:
      consumes:
      - application/json
      description: Update Warehouse
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Warehouse
        in: body
        name: warehouse
        required: true
        schema:
          $ref: '#/definitions/models.Warehouse'
      - description: Product Id
        in: query
        name: productId
        required: true
        type: integer
      - description: DiscountId Id
        in: query
        name: discountId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Update Warehouse
      tags:
      - warehouses
    post:
      consumes:
      - multipart/form-data
      description: Import Warehouse
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: categoryId
        type: integer
      - in: formData
        name: description
        type: string
      - in: formData
        name: discountPercent
        type: integer
      - in: formData
        name: minQuantity
        type: integer
      - in: formData
        name: name
        type: string
      - in: formData
        name: price
        type: number
      - in: formData
        name: stockAmount
        type: integer
      - in: formData
        name: '-'
        type: string
      - in: formData
        name: unit
        type: string
      - description: Thumb
        in: formData
        name: thumb
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Import Warehouse
      tags:
      - warehouses
swagger: "2.0"